remove `showInterface` in favour of `createInterface` and refactor `InterfaceController`
