+++ packages/snaps-rpc-methods/src/restricted/createInterface.ts
